+++ .clang-format
@@ -17,4 +17,3 @@ DerivePointerAlignment: false
 PointerAlignment: Middle
 ReflowComments: false
 ...
-
