+++ CMakeLists.txt
@@ -14,16 +14,16 @@
 # limitations under the License.
 
 cmake_minimum_required(VERSION 3.5)
-project(metavision_ros_driver)
+project(metavision_driver)
 
 if(DEFINED ENV{ROS_VERSION})
   if($ENV{ROS_VERSION} EQUAL "2")
     # to avoid cmake warning: "Manually-specified variables were not used by the project"
     unset(CATKIN_INSTALL_INTO_PREFIX_ROOT)
     unset(CATKIN_SYMLINK_INSTALL)
-    include(${CMAKE_CURRENT_SOURCE_DIR}/cmake/CMakeListsROS2.txt)
+    include(${CMAKE_CURRENT_SOURCE_DIR}/cmake/ROS2.cmake)
   elseif($ENV{ROS_VERSION} EQUAL "1")
-    include(${CMAKE_CURRENT_SOURCE_DIR}/cmake/CMakeListsROS1.txt)
+    include(${CMAKE_CURRENT_SOURCE_DIR}/cmake/ROS1.cmake)
   else()
     message(SEND_ERROR "invalid ROS_VERSION environment variable: $ENV{ROS_VERSION}")
   endif()
