+++ cfg/MetaVisionDyn.cfg
@@ -1,5 +1,5 @@
 #!/usr/bin/env python2
-PACKAGE = "metavision_ros_driver"
+PACKAGE = "metavision_driver"
 
 from dynamic_reconfigure.parameter_generator_catkin import *
 
@@ -57,4 +57,4 @@ gen.add("bias_pr", int_t, 0, "photoreceptor (frontend) bias", 1500, 0, 1800);
 # the pixels and hence the output data rate of the sensor.
 
 gen.add("bias_refr", int_t, 0, "refractory time bias", -20, 0, 1800);
-exit(gen.generate(PACKAGE, "metavision_ros_driver", "MetaVisionDyn"))
+exit(gen.generate(PACKAGE, "metavision_driver", "MetaVisionDyn"))
