+++ include/metavision_driver/callback_handler.h
@@ -13,12 +13,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef METAVISION_ROS_DRIVER__CALLBACK_HANDLER_H_
-#define METAVISION_ROS_DRIVER__CALLBACK_HANDLER_H_
+#ifndef METAVISION_DRIVER__CALLBACK_HANDLER_H_
+#define METAVISION_DRIVER__CALLBACK_HANDLER_H_
 
 #include <metavision/sdk/driver/camera.h>
 
-namespace metavision_ros_driver
+namespace metavision_driver
 {
 class CallbackHandler
 {
@@ -29,5 +29,5 @@ public:
   virtual void eventCDCallback(
     uint64_t t, const Metavision::EventCD * start, const Metavision::EventCD * end) = 0;
 };
-}  // namespace metavision_ros_driver
-#endif  // METAVISION_ROS_DRIVER__CALLBACK_HANDLER_H_
+}  // namespace metavision_driver
+#endif  // METAVISION_DRIVER__CALLBACK_HANDLER_H_
