+++ nodelet_plugins.xml
@@ -1,9 +1,9 @@
 <library path="lib/libdriver_nodelet">
-  <class name="metavision_ros_driver/DriverNodelet"
-         type="metavision_ros_driver::DriverNodeletROS1"
+  <class name="metavision_driver/DriverNodelet"
+         type="metavision_driver::DriverNodeletROS1"
          base_class_type="nodelet::Nodelet">
     <description>
-      metavision driver nodelet
+      ROS metavision driver nodelet
     </description>
   </class>
 </library>
