+++ src/driver_node_ros2.cpp
@@ -16,12 +16,12 @@
 #include <memory>
 #include <rclcpp/rclcpp.hpp>
 
-#include "metavision_ros_driver/driver_ros2.h"
+#include "metavision_driver/driver_ros2.h"
 
 int main(int argc, char * argv[])
 {
   rclcpp::init(argc, argv);
-  auto node = std::make_shared<metavision_ros_driver::DriverROS2>(rclcpp::NodeOptions());
+  auto node = std::make_shared<metavision_driver::DriverROS2>(rclcpp::NodeOptions());
 
   RCLCPP_INFO(node->get_logger(), "driver_node started up!");
   // actually run the node
